crypto-etl-pipeline/requirements.txt
requests>=2.31.0
pandas>=2.2.0
duckdb>=0.10.0
prefect>=2.14.0
pytest>=8.0.0
python-dotenv>=1.0.0
